# api/openapi.yml
openapi: 3.0.3
info:
  title: MTS Online Shop
  version: 0.1.0
  description: REST API интернет-магазина
servers:
  - url: http://localhost:8080/api
    description: Dev Server
...
paths:
...
  /orders:
    get:
      summary: Получить заказы текущего пользователя
      tags: [Orders]
      operationId: getOrders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Список заказов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
  /orders/create:
    post:
      summary: Создать заказ с оплатой (транзакция)
      tags: [Orders]
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
            examples:
              payment_example:
                summary: "Example payment data"
                value:
                  cardNumber: "1234123412341234"
                  cvv: "444"
                  expiresAt: "09/30"
      responses:
        "201":
          description: Заказ создан и оплачен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Ошибка валидации данных карты или пустая корзина
        "500":
          description: Внутренняя ошибка сервера
  /orders/{orderId}/cancel:
    post:
      summary: Отменить заказ с возвратом денег
      tags: [Orders]
      operationId: cancelOrder
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Заказ отменен, деньги возвращены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
                
  # Admin APIs
...
  /admin/orders/{orderId}/cancel:
    post:
      summary: Отменить любой заказ с возвратом
      tags: [Admin Orders]
      operationId: adminCancelOrder
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Заказ отменен, деньги возвращены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Не авторизован
        "403":
          description: Недостаточно прав (только ADMIN)
        "404":
          description: Заказ не найден
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "Операция выполнена успешно"
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Пакет интернета 5Gb"
        price:
          type: number
          format: double
          example: 199.99
...
    CartItem:
      type: object
      properties:
        itemId:
          type: integer
          format: int64
          example: 123
        product:
          $ref: "#/components/schemas/Product"
...
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        userId:
          type: integer
          format: int64
          example: 2
        totalPrice:
          type: number
          format: double
          example: 599.97
        status:
          type: string
          enum: [CREATED, PAID, DELIVERED, CANCELLED]
          example: "PAID"
        createdAt:
          type: string
          format: date-time
          example: "2024-03-15T10:30:00Z"
    OrderResponse:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/CartItem"
    OrderListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderResponse"
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
...
    PaymentRequest:
      type: object
      required:
        - cardNumber
        - expiresAt
        - cvv
      properties:
        cardNumber:
          type: string
          example: "1234123412341234"
        expiresAt:
          type: string
          example: "12/30"
        cvv:
          type: string
          example: "111"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT